backend/requirements.txt
fastapi==0.115.4
uvicorn==0.34.0
requests==2.31.0
transformers==4.44.2
spacy==3.7.4
diffusers==0.29.0
torch==2.4.1
python-multipart==0.0.12
python-dotenv==1.0.1
accelerate==0.34.2